Split integration tests by feature
